requirements.txt
requests
pandas
beautifulsoup4
selenium
webdriver-manager
python-dotenv
openpyxl
Jinja2



